refactor: 기존 테스트 디렉토리 리팩토링
