+++ src/test/java/subway/acceptanceTest/StationAcceptanceTest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.acceptanceTest;
 
 import config.annotations.AcceptanceTest;
 import io.restassured.RestAssured;
